Rename TKClient -> MobileRestClient, add some documentation, rename generators.py to doc_generators.py
