+++ testsuites/listener/shared/replication.robot
@@ -7,7 +7,7 @@ Library           DebugLibrary
 Library           Process
 
 Library           ${KEYWORDS}/Async.py
-Library           ${KEYWORDS}/TKClient.py
+Library           ${KEYWORDS}/MobileRestClient.py
 Library           ${KEYWORDS}/LiteServ.py
 ...                 platform=${PLATFORM}
 ...                 version_build=${LITESERV_VERSION}
